Fix a bug that favorites don't contain ratings
