+++ backend/controllers/marketController.js
@@ -1,21 +1,7 @@
 const Market = require('../models/marketModel');
 const User = require('../models/userModel');
-const Comment = require('../models/commentModel');
 const CookieManager = require("../shared/cookieManager");
-
-// 평점, 리뷰 수 등은 동적으로 계산
-async function getDynamicInfo(market) {
-    const comments = await Comment.find({comment_target: market.market_index});
-    let ratAcc = 0.0;
-    comments.forEach(c => {
-        ratAcc += c.comment_score;
-    });
-    
-    const augmented = market.toObject();
-    augmented.market_rating = comments.length ? ratAcc / comments.length : 0;
-    augmented.market_comments_count = comments.length;
-    return augmented;
-}
+const getDynamicInfo = require("../shared/marketDynamicInfo");
 
 async function GetAllMarkets(req, res) {
     try {
